backend/requirements.txt
fastapi==0.115.6
uvicorn==0.30.6
gunicorn==21.2.0
pydantic==2.10.3
python-dotenv==1.0.1
python-multipart==0.0.9
aiofiles==24.1.0
reportlab==4.2.5
google-cloud-aiplatform==1.68.0
google-cloud-storage==2.18.0
google-auth==2.35.0
Pillow==11.0.0
httpx==0.27.2
sentry-sdk[fastapi]==1.40.0
